修改依赖

+++ Model/Acl.php
@@ -208,7 +208,7 @@ class Acl extends \Weline\Framework\Database\Model
      */
     public function install(ModelSetup $setup, Context $context): void
     {
-        $setup->query('TRUNCATE TABLE ' . $this->getTable());
+        if ($setup->tableExist())$setup->query('TRUNCATE TABLE ' . $this->getTable());
 //        $setup->dropTable();
         if (!$setup->tableExist()) {
             $setup->createTable()
